+++ cache/fast_lru_cache.cc
@@ -30,7 +30,7 @@ LRUHandleTable::LRUHandleTable(int hash_bits)
     : length_bits_(hash_bits),
       length_bits_mask_((uint32_t{1} << length_bits_) - 1),
       occupancy_(0),
-      array_(new LRUHandle[uint32_t{1} << length_bits_]) {
+      array_(new LRUHandle[size_t{1} << length_bits_]) {
   assert(hash_bits <= 32);
 }
